+++ src/utils/has-dom.ts
@@ -1,1 +1,4 @@
-export default () => window && window.document && window.document.createElement;
+export default () =>
+  typeof window !== 'undefined' &&
+  window.document &&
+  window.document.createElement;